Internal function for grabbing and preparing iplot data

Grabs the underlying data used to construct fixest::iplot, with some added functionality and tweaks for the ggiplot equivalents.

Usage

iplot_data(
  object,
  .ci_level = 0.95,
  .dict = fixest::getFixest_dict(),
  .aggr_es = c("none", "post", "pre", "both")
)

Arguments

object

A model object of class fixest or fixest_multi, where the i() operator has been used to construct an interaction, or set of interactions.

.ci_level

A number between 0 and 1 indicating the desired confidence level, Defaults to 0.95.

.dict

A dictionary (i.e. named character vector or a logical scalar). Used for changing coefficient names. Defaults to the values in getFixest_dict(). See the ?fixest::iplot documentation for more information.

.aggr_es

A character string indicating whether the aggregated mean post- (and/or pre-) treatment effect should be added as a column to the returned data frame. Passed to aggr_es(..., aggregation = "mean") and should be one of "none" (the default), "post", "pre", or "both".

Value

A data frame consisting of estimate values, confidence intervals, relative x-axis positions, and other aesthetic information needed to draw a ggplot2 object.

Details

This function is a wrapper around fixest::iplot(..., only.params = TRUE), but with various checks and tweaks to better facilitate plotting with ggplot2 and handling of complex object types (e.g. lists of fixest_multi models)

See also

Examples

library(fixest)

est_did = feols(y ~ x1 + i(period, treat, 5) | id+period,
                data = base_did)
iplot(est_did, only.params = TRUE) # The "base" version
#> $prms
#>     estimate    ci_low   ci_high estimate_names estimate_names_raw is_ref  x id
#> 1  -1.403045 -3.579128 0.7730371              1    period::1:treat  FALSE  1  1
#> 2  -1.247511 -3.390035 0.8950132              2    period::2:treat  FALSE  2  1
#> 3  -0.273206 -2.442759 1.8963469              3    period::3:treat  FALSE  3  1
#> 4  -1.795721 -3.928110 0.3366688              4    period::4:treat  FALSE  4  1
#> 5   0.000000  0.000000 0.0000000              5    period::5:treat   TRUE  5  1
#> 6   0.784452 -1.231152 2.8000556              6    period::6:treat  FALSE  6  1
#> 7   3.598897  1.439873 5.7579220              7    period::7:treat  FALSE  7  1
#> 8   3.811766  1.366707 6.2568251              8    period::8:treat  FALSE  8  1
#> 9   4.731426  2.581125 6.8817275              9    period::9:treat  FALSE  9  1
#> 10  6.606229  4.410100 8.8023573             10   period::10:treat  FALSE 10  1
#>            y
#> 1  -1.403045
#> 2  -1.247511
#> 3  -0.273206
#> 4  -1.795721
#> 5   0.000000
#> 6   0.784452
#> 7   3.598897
#> 8   3.811766
#> 9   4.731426
#> 10  6.606229
#> 
#> $is_iplot
#> [1] TRUE
#> 
#> $at
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $labels
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
iplot_data(est_did)                # The wrapper provided by this package
#>     estimate    ci_low   ci_high estimate_names estimate_names_raw is_ref  x id
#> 1  -1.403045 -3.579128 0.7730371              1    period::1:treat  FALSE  1  1
#> 2  -1.247511 -3.390035 0.8950132              2    period::2:treat  FALSE  2  1
#> 3  -0.273206 -2.442759 1.8963469              3    period::3:treat  FALSE  3  1
#> 4  -1.795721 -3.928110 0.3366688              4    period::4:treat  FALSE  4  1
#> 5   0.000000  0.000000 0.0000000              5    period::5:treat   TRUE  5  1
#> 6   0.784452 -1.231152 2.8000556              6    period::6:treat  FALSE  6  1
#> 7   3.598897  1.439873 5.7579220              7    period::7:treat  FALSE  7  1
#> 8   3.811766  1.366707 6.2568251              8    period::8:treat  FALSE  8  1
#> 9   4.731426  2.581125 6.8817275              9    period::9:treat  FALSE  9  1
#> 10  6.606229  4.410100 8.8023573             10   period::10:treat  FALSE 10  1
#>            y dep_var ci_level
#> 1  -1.403045       y     0.95
#> 2  -1.247511       y     0.95
#> 3  -0.273206       y     0.95
#> 4  -1.795721       y     0.95
#> 5   0.000000       y     0.95
#> 6   0.784452       y     0.95
#> 7   3.598897       y     0.95
#> 8   3.811766       y     0.95
#> 9   4.731426       y     0.95
#> 10  6.606229       y     0.95

# Illustrative fixest_multi case, where the sample has been split by odd and
# even ID numbers.
est_split = feols(y ~ x1 + i(period, treat, 5) | id+period,
                  data = base_did, split = ~id%%2)
iplot(est_split, only.params = TRUE) # The "base" version
#> $prms
#>      estimate     ci_low    ci_high    x id estimate_names estimate_names_raw
#> 1   1.4239576 -1.2422952  4.0902104  0.9  1              1    period::1:treat
#> 2  -4.1291547 -7.3973098 -0.8609996  1.1  2              1    period::1:treat
#> 3   1.6931989 -1.1310149  4.5174127  1.9  1              2    period::2:treat
#> 4  -4.1353019 -7.2460773 -1.0245265  2.1  2              2    period::2:treat
#> 5   3.0439521  0.3506812  5.7372230  2.9  1              3    period::3:treat
#> 6  -3.4897091 -6.5504723 -0.4289459  3.1  2              3    period::3:treat
#> 7  -0.8667871 -3.8315352  2.0979611  3.9  1              4    period::4:treat
#> 8  -2.6458897 -5.7093088  0.4175294  4.1  2              4    period::4:treat
#> 9   0.0000000  0.0000000  0.0000000  4.9  1              5    period::5:treat
#> 10  0.0000000  0.0000000  0.0000000  5.1  2              5    period::5:treat
#> 11  3.8181620  0.9644230  6.6719010  5.9  1              6    period::6:treat
#> 12 -2.2626440 -4.9448774  0.4195895  6.1  2              6    period::6:treat
#> 13  5.1642678  2.1084814  8.2200543  6.9  1              7    period::7:treat
#> 14  2.0828305 -0.9816140  5.1472751  7.1  2              7    period::7:treat
#> 15  5.2338735  2.0448159  8.4229311  7.9  1              8    period::8:treat
#> 16  2.4826394 -1.2053985  6.1706773  8.1  2              8    period::8:treat
#> 17  7.0979736  3.9704919 10.2254553  8.9  1              9    period::9:treat
#> 18  2.4385390 -0.2928821  5.1699601  9.1  2              9    period::9:treat
#> 19  9.5453455  6.3977727 12.6929184  9.9  1             10   period::10:treat
#> 20  3.6267443  0.6409795  6.6125091 10.1  2             10   period::10:treat
#>    is_ref          y
#> 1   FALSE  1.4239576
#> 2   FALSE -4.1291547
#> 3   FALSE  1.6931989
#> 4   FALSE -4.1353019
#> 5   FALSE  3.0439521
#> 6   FALSE -3.4897091
#> 7   FALSE -0.8667871
#> 8   FALSE -2.6458897
#> 9    TRUE  0.0000000
#> 10   TRUE  0.0000000
#> 11  FALSE  3.8181620
#> 12  FALSE -2.2626440
#> 13  FALSE  5.1642678
#> 14  FALSE  2.0828305
#> 15  FALSE  5.2338735
#> 16  FALSE  2.4826394
#> 17  FALSE  7.0979736
#> 18  FALSE  2.4385390
#> 19  FALSE  9.5453455
#> 20  FALSE  3.6267443
#> 
#> $is_iplot
#> [1] TRUE
#> 
#> $at
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $labels
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
iplot_data(est_split)                # The wrapper provided by this package
#>      estimate     ci_low    ci_high  x                           id
#> 1   1.4239576 -1.2422952  4.0902104  1 sample.var: id%%2; sample: 0
#> 2  -4.1291547 -7.3973098 -0.8609996  1 sample.var: id%%2; sample: 1
#> 3   1.6931989 -1.1310149  4.5174127  2 sample.var: id%%2; sample: 0
#> 4  -4.1353019 -7.2460773 -1.0245265  2 sample.var: id%%2; sample: 1
#> 5   3.0439521  0.3506812  5.7372230  3 sample.var: id%%2; sample: 0
#> 6  -3.4897091 -6.5504723 -0.4289459  3 sample.var: id%%2; sample: 1
#> 7  -0.8667871 -3.8315352  2.0979611  4 sample.var: id%%2; sample: 0
#> 8  -2.6458897 -5.7093088  0.4175294  4 sample.var: id%%2; sample: 1
#> 9   0.0000000  0.0000000  0.0000000  5 sample.var: id%%2; sample: 0
#> 10  0.0000000  0.0000000  0.0000000  5 sample.var: id%%2; sample: 1
#> 11  3.8181620  0.9644230  6.6719010  6 sample.var: id%%2; sample: 0
#> 12 -2.2626440 -4.9448774  0.4195895  6 sample.var: id%%2; sample: 1
#> 13  5.1642678  2.1084814  8.2200543  7 sample.var: id%%2; sample: 0
#> 14  2.0828305 -0.9816140  5.1472751  7 sample.var: id%%2; sample: 1
#> 15  5.2338735  2.0448159  8.4229311  8 sample.var: id%%2; sample: 0
#> 16  2.4826394 -1.2053985  6.1706773  8 sample.var: id%%2; sample: 1
#> 17  7.0979736  3.9704919 10.2254553  9 sample.var: id%%2; sample: 0
#> 18  2.4385390 -0.2928821  5.1699601  9 sample.var: id%%2; sample: 1
#> 19  9.5453455  6.3977727 12.6929184 10 sample.var: id%%2; sample: 0
#> 20  3.6267443  0.6409795  6.6125091 10 sample.var: id%%2; sample: 1
#>    estimate_names estimate_names_raw is_ref          y dep_var ci_level
#> 1               1    period::1:treat  FALSE  1.4239576       y     0.95
#> 2               1    period::1:treat  FALSE -4.1291547       y     0.95
#> 3               2    period::2:treat  FALSE  1.6931989       y     0.95
#> 4               2    period::2:treat  FALSE -4.1353019       y     0.95
#> 5               3    period::3:treat  FALSE  3.0439521       y     0.95
#> 6               3    period::3:treat  FALSE -3.4897091       y     0.95
#> 7               4    period::4:treat  FALSE -0.8667871       y     0.95
#> 8               4    period::4:treat  FALSE -2.6458897       y     0.95
#> 9               5    period::5:treat   TRUE  0.0000000       y     0.95
#> 10              5    period::5:treat   TRUE  0.0000000       y     0.95
#> 11              6    period::6:treat  FALSE  3.8181620       y     0.95
#> 12              6    period::6:treat  FALSE -2.2626440       y     0.95
#> 13              7    period::7:treat  FALSE  5.1642678       y     0.95
#> 14              7    period::7:treat  FALSE  2.0828305       y     0.95
#> 15              8    period::8:treat  FALSE  5.2338735       y     0.95
#> 16              8    period::8:treat  FALSE  2.4826394       y     0.95
#> 17              9    period::9:treat  FALSE  7.0979736       y     0.95
#> 18              9    period::9:treat  FALSE  2.4385390       y     0.95
#> 19             10   period::10:treat  FALSE  9.5453455       y     0.95
#> 20             10   period::10:treat  FALSE  3.6267443       y     0.95